# IndiaCP_SwaggerSpecificaiton/src/main/resources/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IndiaCP API",
        "description": "This API will drive the UI",
        "version": "1.0.0"
    },
    "host": "finwizui.azurewebsites.net",
    "schemes": [
        "http"
    ],
    "basePath": "/api",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/indiacpprogram/issueCPProgram": {
            "post": {
                "summary": "Issue new CP Program",
                "description": "This creates a new CP Program with the details provided",
                "operationId": "issueCPProgram",
                "consumes": [ "application/json"],
                "parameters": [
                    {
                        "name": "cpprogramDetails",
                        "in": "body",
                        "description": "Details of the CP Program to be Issued",
                        "required": true,
                        "schema":
                        {
                          "$ref" : "#/definitions/IndiaCPProgram"
                        }
                    }
                ],
                "tags": [
                    "Issuer"
                ],
                "responses": {
                    "200": {
                        "description": "Creates a new CP Program",
                        "schema": {
                            "$ref": "#/definitions/IndiaCPProgram"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/indiacpprogram/fetchAllCPProgram": {
            "get": {
                "summary": "Fetch All CPPrograms for the current Issuer. The current Issuer is assumed from the DL Node that this request is forwarded to.",
                "description": "Returns all the CP Programs for the current issuer\n",
                "operationId": "fetchAllCPProgram",
                "parameters": [
                ],
                "tags": [
                    "Issuer"
                ],
                "responses": {
                    "200": {
                        "description": "An array of CP Programs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/IndiaCPProgram"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/indiacpprogram/fetchCPProgram/{cpProgramId}": {
            "get": {
                "summary": "Get CP Program details by Id. The current Issuer is assumed from the DL Node that this request is forwarded to.",
                "description": "This returns a single CP Program identified by an Id provided by the call\n",
                "operationId": "fetchCPProgram",
                "parameters": [
                    {
                        "name": "cpProgramId",
                        "in": "path",
                        "description": "CP Program ID that uniquely identifies the CP Program issued by the Issuer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Issuer"
                ],
                "responses": {
                    "200": {
                        "description": "Returns a CP Program",
                        "schema": {
                            "$ref": "#/definitions/IndiaCPProgram"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/indiacpprogram/addDocs/{cpProgramId}": {
            "post": {
                "summary": "Uploads and attaches the provided document to the CPProgram on the DL",
                "description": "Uploads and attaches the provided document to the CPProgram on the DL",
                "operationId": "addDoc",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                  "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "cpProgramId",
                        "in": "path",
                        "description": "CP Program ID that uniquely identifies the CP Program issued by the Issuer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "metadata",
                        "in": "formData",
                        "description": "Document Attachment Details",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Document Attachment",
                        "required": true,
                        "type": "file"
                    }
                ],
                "tags": [
                    "Issuer"
                ],
                "responses": {
                    "200": {
                        "description": "Returns a CP Program",
                        "schema": {
                            "$ref": "#/definitions/IndiaCPProgram"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/indiacpprogram/addISIN/{cpProgramId}/{isin}": {
            "post": {
                "summary": "Adds the ISIN to the India CP Program",
                "description": "Adds the ISIN to the India CP Program",
                "operationId": "addISIN",
                "parameters": [
                    {
                        "name": "cpProgramId",
                        "in": "path",
                        "description": "CP Program ID that uniquely identifies the CP Program issued by the Issuer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "isin",
                        "in": "path",
                        "description": "ISIN generated by NSDL",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Issuer"
                ],
                "responses": {
                    "200": {
                        "description": "Returns a CP Program",
                        "schema": {
                            "$ref": "#/definitions/IndiaCPProgram"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/indiacpprogram/getTransactionHistory/{cpProgramId}": {
            "get": {
                "summary": "Complete audit log of all changes/versions of given CP Program",
                "description": "A given CP Program once initiated undergoes various changes as it progresses through the trade lifecycle of generating ISIN, generating Deal Confirmations with each identified Investor, getting IPA Verification till the final settlement of all Deals and followed by redemption of the CP at the Maturity Date. This API will return this complete log history.",
                "parameters": [
                    {
                        "name": "cpProgramId",
                        "in": "path",
                        "description": "CP Program ID that uniquely identifies the CP Program issued by the Issuer",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Issuer, Investor, IPA, Depository"
                ],
                "responses": {
                    "200": {
                        "description": "An array of all versions of the given CP Program",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/IndiaCPProgram"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/indiacpissue/addDocs/{cpIssueId}": {
            "post": {
                "summary": "Uploads and attaches the provided document to the CP on the DL",
                "description": "Uploads and attaches the provided document to the CP on the DL",
                "operationId": "addDocs",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "cpIssueId",
                        "in": "path",
                        "description": "Unique identifier of the CP Issue",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "metadata",
                        "in": "formData",
                        "description": "Document Attachment Details",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Document Attachment",
                        "required": true,
                        "type": "file"
                    }
                ],
                "tags": [
                    "Issuer, Investor, IPA, Depository"
                ],
                "responses": {
                    "200": {
                        "description": "Returns a CP Issue",
                        "schema": {
                            "$ref": "#/definitions/IndiaCPIssue"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "IndiaCPProgram": {
            "type": "object",
            "properties": {
                "programId": {
                    "type": "string",
                    "description": "Unique identifier representing a specific CP Program raised by an Issuer"
                },
                "name": {
                    "type": "string",
                    "description": "Name of the CP Program"
                },
                "type": {
                    "type": "string",
                    "description": "Type of the CP Program"
                },
                "purpose": {
                    "type": "string",
                    "description": "Purpose of the CP Program"
                },
                "isin": {
                    "type": "string",
                    "description": "Unique CP Security Identifier No. In India this is issued by NSDL for Commercial Paper type of securities."
                },
                "issuerId": {
                    "type": "string",
                    "description": "Unique identifier of the Issuer. This will also uniquely map to the DL Node of the Issuer"
                },
                "issuerName": {
                    "type": "string",
                    "description": "Display name of the Issuer"
                },
                "issueCommencementDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Date when the CP Program was initiated. India regulations mandate that a program should be fully allocated within 2 weeks of commencement."
                },
                "programSize": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Total borrowing amount under this CP Program"
                },
                "programAllocatedValue": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Amount already allocated to the CP program"
                },
                "programCurrency": {
                    "type": "string",
                    "description": "Currency of the issue of CP notes"
                },
                "maturityDays": {
                    "type": "integer",
                    "description": "No of days to maturity from the value date"
                },
                "ipaId": {
                    "type": "string",
                    "description": "Unique identifier of the IPA. This will also uniquely map to the DL Node of the IPA"
                },
                "ipaName": {
                    "type": "string",
                    "description": "Display name of the IPA"
                },
                "depositoryId": {
                    "type": "string",
                    "description": "Unique identifier of the Depository (NSDL). This will also uniquely map to the DL Node of the Depository"
                },
                "depositoryName": {
                    "type": "string",
                    "description": "Display name of the Depository"
                },
                "isinGenerationRequestDocId": {
                    "type": "string",
                    "description": "Unique identifier of the documents sent to the depository to generate ISIN"
                },
                "ipaVerificationRequestDocId": {
                    "type": "string",
                    "description": "Unique identifier of the documents sent to the IPA to certify the CP Program"
                },
                "ipaCertificateDocId": {
                    "type": "string",
                    "description": "Unique identifier of the IPA certificate issued by the IPA on verification of the CP Program and supporting Issuer documents"
                },
                "corporateActionFormDocId": {
                    "type": "string",
                    "description": "Unique identifier of the Corporate Action Form generated by the Issuer to allot CP"
                },
                "allotmentLetterDocId": {
                    "type": "string",
                    "description": "Unique identifier of the Allotment Letter generated by IPA for CP transfer to Investor DP account"
                },
                "status": {
                    "type": "string",
                    "description": "Current status of the CP Program"
                },
                "version": {
                    "type": "integer",
                    "description": "Current version of the CP Program"
                },
                "modifiedBy": {
                    "type": "string",
                    "description": "Unique identifier of the Logged-in User that performed the action. This is required for Audit History"
                },
                "lastModifiedDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Last Modified Date for this CPIssue. This is required for Audit History"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "description": "Unique Id to identify the type of error. For e.g. CP Program Creation, ISIN Generation, etc."
                },
                "source": {
                    "type": "string",
                    "description": "Source of the error identifying the layer where the error originated. For e.g. DL Integration Layer, DL (Corda/Ethereum), Reference Data Service",
                    "enum": [
                        "DL_INTEGRATION_LAYER",
                        "DL_R3CORDA",
                        "DL_ETHEREUM_QUORUM_CAKESHOP",
                        "DL_ETHEREUM_QUORUM",
                        "REFERENCE_DATA_SERVICE",
                        "UNKNOWN_SOURCE"
                    ]
                },
                "message": {
                    "type": "string",
                    "description": "Short message describing the error"
                },
                "details": {
                    "type": "string",
                    "description": "Details of the error. This could also be the complete stack trace. It would be useful for debugging"
                }
            }
        },
...
    }
}